Function "calc_AO_overlap" name updated to "calc_AO_overlap_from_xyz".
  Function "get_AO_overlap" added to read atomic orbital overlaps from
NWChem standard output.

  Fortran subroutines added to read molecular orbital vectors written
in binary from NWChem with extensions ".movecs".

  Notes on the molecular orbital files from NWWChem. The molecular
orbitals written in the ".molden" have some non-trivial swaps for
compatibility purpurses with other software and are not consisten with
the atomic orbital overlaps written in the standard output. In order
to use molecular orbitals and atomic orbital overlaps consistent with
each other, it is better to read the basis information from the
".molden" files and calculate the atomic orbital overlaps with this
information.